+++ packages/web3-eth-accounts/src/account.ts
@@ -33,7 +33,15 @@ import {
 	TransactionSigningError,
 	UndefinedRawTransactionError,
 } from 'web3-errors';
-import { Address, Bytes, HexString } from 'web3-types';
+import {
+	Address,
+	Bytes,
+	HexString,
+	CipherOptions,
+	PBKDF2SHA256Params,
+	ScryptParams,
+	KeyStore,
+} from 'web3-types';
 import {
 	bytesToBuffer,
 	bytesToHex,
@@ -44,19 +52,11 @@ import {
 	sha3Raw,
 	toChecksumAddress,
 	utf8ToHex,
+	uuidV4,
 } from 'web3-utils';
 import { isBuffer, isNullish, isString, validator } from 'web3-validator';
 import { keyStoreSchema } from './schemas';
-import {
-	CipherOptions,
-	KeyStore,
-	PBKDF2SHA256Params,
-	ScryptParams,
-	SignatureObject,
-	SignResult,
-	SignTransactionResult,
-	Web3Account,
-} from './types';
+import { SignatureObject, SignResult, SignTransactionResult, Web3Account } from './types';
 
 /**
  * Get the private key buffer after the validation
@@ -355,37 +355,6 @@ export const recover = (
 };
 
 /**
- * Generate a version 4 (random) uuid
- * https://github.com/uuidjs/uuid/blob/main/src/v4.js#L5
- */
-
-const uuidV4 = (): string => {
-	const bytes = randomBytes(16);
-
-	// https://github.com/ethers-io/ethers.js/blob/ce8f1e4015c0f27bf178238770b1325136e3351a/packages/json-wallets/src.ts/utils.ts#L54
-	// Section: 4.1.3:
-	// - time_hi_and_version[12:16] = 0b0100
-	/* eslint-disable-next-line */
-	bytes[6] = (bytes[6] & 0x0f) | 0x40;
-
-	// Section 4.4
-	// - clock_seq_hi_and_reserved[6] = 0b0
-	// - clock_seq_hi_and_reserved[7] = 0b1
-	/* eslint-disable-next-line */
-	bytes[8] = (bytes[8] & 0x3f) | 0x80;
-
-	const hexString = bytesToHex(bytes);
-
-	return [
-		hexString.substring(2, 10),
-		hexString.substring(10, 14),
-		hexString.substring(14, 18),
-		hexString.substring(18, 22),
-		hexString.substring(22, 34),
-	].join('-');
-};
-
-/**
  * Get the ethereum Address from a private key
  *
  * @param privateKey - String or buffer of 32 bytes
@@ -574,7 +543,6 @@ export const encrypt = async (
 	const ciphertext = bytesToHex(cipher).slice(2);
 
 	const mac = sha3Raw(Buffer.from([...derivedKey.slice(16, 32), ...cipher])).replace('0x', '');
-
 	return {
 		version: 3,
 		id: uuidV4(),
@@ -625,11 +593,8 @@ export const privateKeyToAccount = (privateKey: Bytes, ignoreLength?: boolean): 
 		},
 		sign: (data: Record<string, unknown> | string) =>
 			sign(typeof data === 'string' ? data : JSON.stringify(data), privateKeyBuffer),
-		encrypt: async (password: string, options?: Record<string, unknown>) => {
-			const data = await encrypt(privateKeyBuffer, password, options);
-
-			return JSON.stringify(data);
-		},
+		encrypt: async (password: string, options?: Record<string, unknown>) =>
+			encrypt(privateKeyBuffer, password, options),
 	};
 };