+++ packages/web3-types/src/web3_base_provider.ts
@@ -70,9 +70,8 @@ export interface LegacyRequestProvider {
 
 export interface EIP1193Provider<API extends Web3APISpec> {
 	request<Method extends Web3APIMethod<API>, ResponseType = Web3APIReturnType<API, Method>>(
-		request: Web3APIPayload<API, Method>,
-		requestOptions?: unknown,
-	): Promise<JsonRpcResponseWithResult<ResponseType>>;
+		args: Web3APIPayload<API, Method>,
+	): Promise<JsonRpcResponseWithResult<ResponseType> | unknown>;
 }
 
 // Provider interface compatible with EIP-1193
@@ -104,13 +103,14 @@ export abstract class Web3BaseProvider<API extends Web3APISpec = EthExecutionAPI
 	 * @param payload - Request Payload
 	 * @param callback - Callback
 	 */
-	public send<R = JsonRpcResult, P = unknown>(
+	public send<ResultType = JsonRpcResult, P = unknown>(
 		payload: JsonRpcPayload<P>,
 		// eslint-disable-next-line @typescript-eslint/ban-types
-		callback: (err?: Error | null, response?: JsonRpcResponse<R>) => void,
+		callback: (err?: Error | null, response?: JsonRpcResponse<ResultType>) => void,
 	) {
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		this.request(payload as Web3APIPayload<API, any>)
+		this.request<Web3APIMethod<API>, ResultType>(
+			payload as Web3APIPayload<API, Web3APIMethod<API>>,
+		)
 			.then(response => {
 				callback(undefined, response);
 			})
@@ -124,18 +124,16 @@ export abstract class Web3BaseProvider<API extends Web3APISpec = EthExecutionAPI
 	 * @param payload - Request Payload
 	 */
 	public async sendAsync<R = JsonRpcResult, P = unknown>(payload: JsonRpcPayload<P>) {
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		return this.request(payload as Web3APIPayload<API, any>) as Promise<JsonRpcResponse<R>>;
+		return this.request(payload as Web3APIPayload<API, Web3APIMethod<API>>) as Promise<
+			JsonRpcResponse<R>
+		>;
 	}
 
 	// https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1193.md#request
 	public abstract request<
 		Method extends Web3APIMethod<API>,
-		ResultType = Web3APIReturnType<API, Method>,
-	>(
-		request: Web3APIPayload<API, Method>,
-		requestOptions?: unknown,
-	): Promise<JsonRpcResponseWithResult<ResultType>>;
+		ResultType = Web3APIReturnType<API, Method> | unknown,
+	>(args: Web3APIPayload<API, Method>): Promise<JsonRpcResponseWithResult<ResultType>>;
 
 	// https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1193.md#events
 	public abstract on<T = JsonRpcResult>(
@@ -169,8 +167,8 @@ export abstract class Web3BaseProvider<API extends Web3APISpec = EthExecutionAPI
 }
 
 export type SupportedProviders<API extends Web3APISpec> =
-	| Web3BaseProvider<API>
 	| EIP1193Provider<API>
+	| Web3BaseProvider<API>
 	| LegacyRequestProvider
 	| LegacySendProvider
 	| LegacySendAsyncProvider;