+++ packages/web3-types/src/web3_base_wallet.ts
@@ -16,7 +16,48 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { HexString } from './primitives_types';
 
-export type Web3EncryptedWallet = string;
+export type Cipher = 'aes-128-ctr' | 'aes-128-cbc' | 'aes-256-cbc';
+
+export type CipherOptions = {
+	salt?: Buffer | string;
+	iv?: Buffer | string;
+	kdf?: 'scrypt' | 'pbkdf2';
+	dklen?: number;
+	c?: number; // iterrations
+	n?: number; // cpu/memory cost
+	r?: number; // block size
+	p?: number; // parallelization cost
+};
+
+export type ScryptParams = {
+	dklen: number;
+	n: number;
+	p: number;
+	r: number;
+	salt: Buffer | string;
+};
+export type PBKDF2SHA256Params = {
+	c: number; // iterations
+	dklen: number;
+	prf: 'hmac-sha256';
+	salt: Buffer | string;
+};
+
+export type KeyStore = {
+	crypto: {
+		cipher: Cipher;
+		ciphertext: string;
+		cipherparams: {
+			iv: string;
+		};
+		kdf: 'pbkdf2' | 'scrypt';
+		kdfparams: ScryptParams | PBKDF2SHA256Params;
+		mac: HexString;
+	};
+	id: string;
+	version: 3;
+	address: string;
+};
 
 export interface Web3BaseWalletAccount {
 	[key: string]: unknown;
@@ -38,16 +79,17 @@ export interface Web3BaseWalletAccount {
 		readonly message?: string;
 		readonly signature: HexString;
 	};
-	readonly encrypt: (
-		password: string,
-		options?: Record<string, unknown>,
-	) => Promise<Web3EncryptedWallet>;
+	readonly encrypt: (password: string, options?: Record<string, unknown>) => Promise<KeyStore>;
 }
 
 export interface Web3AccountProvider<T> {
 	privateKeyToAccount: (privateKey: string) => T;
 	create: () => T;
-	decrypt: (keystore: string, password: string, options?: Record<string, unknown>) => Promise<T>;
+	decrypt: (
+		keystore: KeyStore | string,
+		password: string,
+		options?: Record<string, unknown>,
+	) => Promise<T>;
 }
 
 export abstract class Web3BaseWallet<T extends Web3BaseWalletAccount> extends Array<T> {
@@ -66,9 +108,9 @@ export abstract class Web3BaseWallet<T extends Web3BaseWalletAccount> extends Ar
 	public abstract encrypt(
 		password: string,
 		options?: Record<string, unknown>,
-	): Promise<Web3EncryptedWallet[]>;
+	): Promise<KeyStore[]>;
 	public abstract decrypt(
-		encryptedWallet: Web3EncryptedWallet[],
+		encryptedWallet: KeyStore[],
 		password: string,
 		options?: Record<string, unknown>,
 	): Promise<this>;