+++ packages/web3-utils/src/converters.ts
@@ -20,6 +20,7 @@ import {
 	validator,
 	isAddress,
 	isHexStrict,
+	isHex,
 	utils as validatorUtils,
 	isNullish,
 } from 'web3-validator';
@@ -214,7 +215,7 @@ export const toAscii = hexToAscii;
  * Auto converts any given value into it's hex representation.
  */
 export const toHex = (
-	value: Numbers | Bytes | Address | boolean,
+	value: Numbers | Bytes | Address | boolean | object,
 	returnType?: boolean,
 ): HexString | ValueTypes => {
 	if (typeof value === 'string' && isAddress(value)) {
@@ -233,6 +234,10 @@ export const toHex = (
 
 	if (typeof value === 'bigint') {
 		return returnType ? 'bigint' : numberToHex(value);
+	}
+
+	if (typeof value === 'object' && !!value) {
+		return returnType ? 'string' : utf8ToHex(JSON.stringify(value));
 	}
 
 	if (typeof value === 'string') {
@@ -267,7 +272,15 @@ export const toNumber = (value: Numbers): number | bigint => {
 			: value;
 	}
 
-	return hexToNumber(numberToHex(value));
+	if (typeof value === 'string' && isHexStrict(value)) {
+		return hexToNumber(value);
+	}
+
+	try {
+		return toNumber(BigInt(value));
+	} catch {
+		throw new InvalidNumberError(value);
+	}
 };
 
 /**
@@ -282,12 +295,9 @@ export const toBigInt = (value: unknown): bigint => {
 		return value;
 	}
 
-	if (typeof value === 'string' && !isHexStrict(value)) {
+	// isHex passes for dec, too
+	if (typeof value === 'string' && isHex(value)) {
 		return BigInt(value);
-	}
-
-	if (typeof value === 'string' && isHexStrict(value)) {
-		return BigInt(hexToNumber(value));
 	}
 
 	throw new InvalidNumberError(value);