+++ packages/web3-types/CHANGELOG.md
@@ -48,3 +48,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `Web3APISpec`, `Web3APIMethod`, and `Web3APIParams` now supports `unknown` APIs (#5393)
 
 ## [Unreleased]
+
+### Added
+
+-   These types were moved from `web3-eth-accounts` to `web3-types` package: Cipher, CipherOptions, ScryptParams, PBKDF2SHA256Params, KeyStore (#5581 )
+
+### Fixed
+
+-   Make the `request` method of `EIP1193Provider` class, compatible with EIP 1193 (#5591)